Updated the kron function to use JAX implementation and moved tests to the nlinalg file
